+++ webshop-api/src/main/java/de/hska/iwi/vslab/webshopapi/models/Category.java
@@ -4,7 +4,7 @@ public class Category {
 
     private Long id;
     private String name;
-    private int[] products;
+    private Long[] products;
 
     public Long getId() {
         return this.id;
@@ -22,11 +22,11 @@ public class Category {
         this.name = name;
     }
 
-    public int[] getProducts() {
+    public Long[] getProducts() {
         return this.products;
     }
 
-    public void setProducts(int[] products) {
+    public void setProducts(Long[] products) {
         this.products = products;
     }
 }
